Refactoring, stage 1 - reduce abconfig clutter
* Move classes and functions out of abconfig.py
  into their own subpackages, keeping related
  definitions together

* Add builders and schedulers properties to the
  Layer() and Distro() classes
